+++ tests/test_functions.py
@@ -10,6 +10,7 @@ import toml
 from ghapi.core import GhApi
 
 from jupyter_releaser import changelog, npm, util
+from jupyter_releaser.changelog import DEFAULT_IGNORED_CONTRIBUTORS
 from jupyter_releaser.util import run
 from tests import util as testutil
 
@@ -70,9 +71,10 @@ def test_get_changelog_version_entry(py_package, mocker):
         since="v1.0",
         until=None,
         kind="pr",
-        branch=branch,
         heading_level=2,
         auth=None,
+        branch=branch,
+        ignored_contributors=DEFAULT_IGNORED_CONTRIBUTORS,
     )
 
     assert f"## {version}" in resp
@@ -87,9 +89,10 @@ def test_get_changelog_version_entry(py_package, mocker):
         since="v1.0",
         until=None,
         kind="pr",
-        branch=branch,
         heading_level=2,
         auth="bizz",
+        branch=branch,
+        ignored_contributors=DEFAULT_IGNORED_CONTRIBUTORS,
     )
 
     assert f"## {version}" in resp
@@ -111,9 +114,10 @@ def test_get_changelog_version_entry_no_tag(py_package, mocker):
         since=commit,
         until=None,
         kind="pr",
-        branch=branch,
         heading_level=2,
         auth=None,
+        branch=branch,
+        ignored_contributors=DEFAULT_IGNORED_CONTRIBUTORS,
     )
 
     assert f"## {version}" in resp
@@ -128,9 +132,10 @@ def test_get_changelog_version_entry_no_tag(py_package, mocker):
         since=commit,
         until=None,
         kind="pr",
-        branch=branch,
         heading_level=2,
         auth="bizz",
+        branch=branch,
+        ignored_contributors=DEFAULT_IGNORED_CONTRIBUTORS,
     )
 
     assert f"## {version}" in resp
@@ -153,9 +158,10 @@ def test_get_changelog_version_entry_since_last_stable(py_package, mocker):
         since="v1.0.0",
         until=None,
         kind="pr",
-        branch=branch,
         heading_level=2,
         auth=None,
+        branch=branch,
+        ignored_contributors=DEFAULT_IGNORED_CONTRIBUTORS,
     )
 
     assert f"## {version}" in resp
@@ -174,12 +180,36 @@ def test_get_empty_changelog(py_package, mocker):
         since="v0.2.4",
         until=None,
         kind="pr",
-        branch=branch,
         heading_level=2,
         auth=None,
+        branch=branch,
+        ignored_contributors=DEFAULT_IGNORED_CONTRIBUTORS,
     )
 
     assert "...None" not in resp
+
+
+def test_get_changelog_no_activity_error(py_package, mocker):
+    """Test handling of ValueError raised by github-activity >= 1.1.4 when no activity is found."""
+    mocked_gen = mocker.patch("jupyter_releaser.changelog.generate_activity_md")
+    mocked_gen.side_effect = ValueError("No activity found for baz/bar between v0.2.4 and None.")
+    branch = "foo"
+    util.run("git branch baz/bar")
+    ref = "heads/baz/bar"
+    resp = changelog.get_version_entry(ref, branch, "baz/bar", "0.2.5", since="v0.2.4")
+    mocked_gen.assert_called_with(
+        "baz/bar",
+        since="v0.2.4",
+        until=None,
+        kind="pr",
+        heading_level=2,
+        auth=None,
+        branch=branch,
+        ignored_contributors=DEFAULT_IGNORED_CONTRIBUTORS,
+    )
+
+    assert "## 0.2.5" in resp
+    assert "No merged PRs" in resp
 
 
 def test_splice_github_entry(py_package, mocker):
@@ -196,12 +226,12 @@ def test_splice_github_entry(py_package, mocker):
 
     updated = changelog.splice_github_entry(resp, testutil.GITHUB_CHANGELOG_ENTRY)
 
-    assert "Defining contributions" in updated
+    assert "Add blog post link" in updated
 
     preamble = "# My title\nmy content\n"
     updated = changelog.splice_github_entry(resp, preamble + testutil.GITHUB_CHANGELOG_ENTRY)
 
-    assert "Defining contributions" in updated
+    assert "Add blog post link" in updated
     assert preamble in updated