+++ CHANGELOG.md
@@ -2,6 +2,51 @@
 
 <!-- <START NEW CHANGELOG ENTRY> -->
 
+## 1.10.2
+
+([Full Changelog](https://github.com/jupyter-server/jupyter_releaser/compare/v2...a79e34830cc24c4051b058402e06284434f721ad))
+
+### Bugs fixed
+
+- Handle `github-activity>=1.1.4` [#637](https://github.com/jupyter-server/jupyter_releaser/pull/637) ([@jtpio](https://github.com/jtpio), [@krassowski](https://github.com/krassowski))
+- Add `meeseeksmachine*` to ignored contributors list [#635](https://github.com/jupyter-server/jupyter_releaser/pull/635) ([@jtpio](https://github.com/jtpio), [@krassowski](https://github.com/krassowski))
+
+### Maintenance and upkeep improvements
+
+- Pin `ruamel.yaml<0.19` in tests [#640](https://github.com/jupyter-server/jupyter_releaser/pull/640) ([@jtpio](https://github.com/jtpio))
+
+### Contributors to this release
+
+The following people contributed discussions, new ideas, code and documentation contributions, and review.
+See [our definition of contributors](https://github-activity.readthedocs.io/en/latest/#how-does-this-tool-define-contributions-in-the-reports).
+
+([GitHub contributors page for this release](https://github.com/jupyter-server/jupyter_releaser/graphs/contributors?from=2025-12-12&to=2026-01-05&type=c))
+
+@jtpio ([activity](https://github.com/search?q=repo%3Ajupyter-server%2Fjupyter_releaser+involves%3Ajtpio+updated%3A2025-12-12..2026-01-05&type=Issues)) | @krassowski ([activity](https://github.com/search?q=repo%3Ajupyter-server%2Fjupyter_releaser+involves%3Akrassowski+updated%3A2025-12-12..2026-01-05&type=Issues))
+
+<!-- <END NEW CHANGELOG ENTRY> -->
+
+## 1.10.1
+
+([Full Changelog](https://github.com/jupyter-server/jupyter_releaser/compare/v2...12393a49f844eca7bc4d08f4db1ee398c2c85605))
+
+### Bugs fixed
+
+- Fix handling of bot accounts [#633](https://github.com/jupyter-server/jupyter_releaser/pull/633) ([@jtpio](https://github.com/jtpio), [@krassowski](https://github.com/krassowski))
+
+### Maintenance and upkeep improvements
+
+- Update changelog test data [#634](https://github.com/jupyter-server/jupyter_releaser/pull/634) ([@jtpio](https://github.com/jtpio), [@krassowski](https://github.com/krassowski))
+
+### Contributors to this release
+
+The following people contributed discussions, new ideas, code and documentation contributions, and review.
+See [our definition of contributors](https://github-activity.readthedocs.io/en/latest/#how-does-this-tool-define-contributions-in-the-reports).
+
+([GitHub contributors page for this release](https://github.com/jupyter-server/jupyter_releaser/graphs/contributors?from=2025-12-11&to=2025-12-12&type=c))
+
+@jtpio ([activity](https://github.com/search?q=repo%3Ajupyter-server%2Fjupyter_releaser+involves%3Ajtpio+updated%3A2025-12-11..2025-12-12&type=Issues)) | @krassowski ([activity](https://github.com/search?q=repo%3Ajupyter-server%2Fjupyter_releaser+involves%3Akrassowski+updated%3A2025-12-11..2025-12-12&type=Issues))
+
 ## 1.10.0
 
 ([Full Changelog](https://github.com/jupyter-server/jupyter_releaser/compare/v2...003a48c40ce52d77a838cb5b2a3b7210c042d003))
@@ -22,8 +67,6 @@ See [our definition of contributors](https://github-activity.readthedocs.io/en/l
 ([GitHub contributors page for this release](https://github.com/jupyter-server/jupyter_releaser/graphs/contributors?from=2025-12-01&to=2025-12-11&type=c))
 
 @brichet ([activity](https://github.com/search?q=repo%3Ajupyter-server%2Fjupyter_releaser+involves%3Abrichet+updated%3A2025-12-01..2025-12-11&type=Issues)) | @jtpio ([activity](https://github.com/search?q=repo%3Ajupyter-server%2Fjupyter_releaser+involves%3Ajtpio+updated%3A2025-12-01..2025-12-11&type=Issues)) | @krassowski ([activity](https://github.com/search?q=repo%3Ajupyter-server%2Fjupyter_releaser+involves%3Akrassowski+updated%3A2025-12-01..2025-12-11&type=Issues)) | @pre-commit-ci\[bot\] ([activity](https://github.com/search?q=repo%3Ajupyter-server%2Fjupyter_releaser+involves%3Apre-commit-ci%5Bbot%5D+updated%3A2025-12-01..2025-12-11&type=Issues))
-
-<!-- <END NEW CHANGELOG ENTRY> -->
 
 ## 1.9.0